move to indexbundle
