+++ composer.json
@@ -140,7 +140,8 @@
     "jms/serializer-bundle": "4.1.0"
   },
   "suggest": {
-    "dpfaffenbauer/process-manager": "Allows to start Processes from within Pimcore UI and also tracks the status."
+    "dpfaffenbauer/process-manager": "Allows to start Processes from within Pimcore UI and also tracks the status.",
+    "elasticsearch/elasticsearch": "Allows to use Elasticsearch as Indexing Engine."
   },
   "autoload": {
     "psr-4": {
@@ -179,7 +180,8 @@
   "config": {
     "sort-packages": true,
     "allow-plugins": {
-      "phpstan/extension-installer": false
+      "phpstan/extension-installer": false,
+      "php-http/discovery": false
     }
   }
 }
