--- src/CoreShop/Bundle/IndexElasticBundle/ElasticsearchBundle.php
@@ -1,91 +1,0 @@
-<?php
-/**
- * CoreShop.
- *
- * This source file is subject to the GNU General Public License version 3 (GPLv3)
- * For the full copyright and license information, please view the LICENSE.md and gpl-3.0.txt
- * files that are distributed with this source code.
- *
- * @copyright  Copyright (c) CoreShop GmbH (https://www.coreshop.org)
- * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
- */
-
-declare(strict_types=1);
-
-namespace CoreShop\Bundle\IndexElasticBundle;
-
-use Composer\InstalledVersions;
-use CoreShop\Bundle\CoreBundle\Application\Version;
-use CoreShop\Bundle\IndexBundle\CoreShopIndexBundle;
-use Pimcore\Extension\Bundle\PimcoreBundleInterface;
-use Pimcore\HttpKernel\BundleCollection\BundleCollection;
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-final class ElasticsearchBundle extends Bundle implements PimcoreBundleInterface
-{
-    public static function registerDependentBundles(BundleCollection $collection): void
-    {
-        $collection->addBundle(new CoreShopIndexBundle(), 3000);
-    }
-
-    public function getNiceName(): string
-    {
-        return 'CoreShop - Elasticsearch';
-    }
-
-    public function getDescription(): string
-    {
-        return 'CoreShop - Elasticsearch Bundle';
-    }
-
-    public function getVersion(): string
-    {
-        $bundleName = 'coreshop/index-elastic-bundle';
-
-        if (class_exists(InstalledVersions::class)) {
-            if (InstalledVersions::isInstalled('coreshop/core-shop')) {
-                return InstalledVersions::getVersion('coreshop/core-shop');
-            }
-
-            if (InstalledVersions::isInstalled($bundleName)) {
-                return InstalledVersions::getVersion($bundleName);
-            }
-        }
-
-        if (class_exists(Version::class)) {
-            return Version::getVersion();
-        }
-
-        return '';
-    }
-
-    public function getInstaller()
-    {
-        return null;
-    }
-
-    public function getAdminIframePath(): ?string
-    {
-        return null;
-    }
-
-    public function getJsPaths(): array
-    {
-        return [];
-    }
-
-    public function getCssPaths(): array
-    {
-        return [];
-    }
-
-    public function getEditmodeJsPaths(): array
-    {
-        return [];
-    }
-
-    public function getEditmodeCssPaths(): array
-    {
-        return [];
-    }
-}
