--- src/CoreShop/Bundle/IndexElasticBundle/Form/Type/Worker/ElasticsearchWorkerTableIndexType.php
@@ -1,52 +1,0 @@
-<?php
-/**
- * CoreShop.
- *
- * This source file is subject to the GNU General Public License version 3 (GPLv3)
- * For the full copyright and license information, please view the LICENSE.md and gpl-3.0.txt
- * files that are distributed with this source code.
- *
- * @copyright  Copyright (c) CoreShop GmbH (https://www.coreshop.org)
- * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
- */
-
-declare(strict_types=1);
-
-namespace CoreShop\Bundle\IndexElasticBundle\Form\Type\Worker;
-
-use CoreShop\Bundle\IndexBundle\Worker\MysqlWorker\TableIndex;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
-use Symfony\Component\Form\Extension\Core\Type\CollectionType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-final class ElasticsearchWorkerTableIndexType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('type', ChoiceType::class, [
-                'choices' => [
-                    TableIndex::TABLE_INDEX_TYPE_INDEX,
-                    TableIndex::TABLE_INDEX_TYPE_UNIQUE,
-                ],
-            ])
-            ->add('columns', CollectionType::class, [
-                'allow_delete' => true,
-                'allow_add' => true,
-            ]);
-    }
-
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver->setDefault('data_class', TableIndex::class);
-
-        parent::configureOptions($resolver);
-    }
-
-    public function getBlockPrefix(): string
-    {
-        return 'coreshop_index_worker_elasticsearch';
-    }
-}
