--- src/CoreShop/Bundle/IndexElasticBundle/Form/Type/Worker/ElasticsearchWorkerType.php
@@ -1,48 +1,0 @@
-<?php
-/**
- * CoreShop.
- *
- * This source file is subject to the GNU General Public License version 3 (GPLv3)
- * For the full copyright and license information, please view the LICENSE.md and gpl-3.0.txt
- * files that are distributed with this source code.
- *
- * @copyright  Copyright (c) CoreShop GmbH (https://www.coreshop.org)
- * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
- */
-
-declare(strict_types=1);
-
-namespace CoreShop\Bundle\IndexElasticBundle\Form\Type\Worker;
-
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\CollectionType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Form\FormBuilderInterface;
-
-final class ElasticsearchWorkerType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('indexes', CollectionType::class, [
-                'entry_type' => ElasticsearchWorkerTableIndexType::class,
-                'allow_add' => true,
-                'allow_delete' => true,
-                'by_reference' => false,
-            ])
-            ->add('localizedIndexes', CollectionType::class, [
-                'entry_type' => ElasticsearchWorkerTableIndexType::class,
-                'allow_add' => true,
-                'allow_delete' => true,
-                'by_reference' => false,
-            ])
-            ->add('hosts', TextType::class, [
-                'label' => 'hosts',
-            ]);
-    }
-
-    public function getBlockPrefix(): string
-    {
-        return 'coreshop_index_worker_elasticsearch';
-    }
-}
