--- src/CoreShop/Bundle/IndexElasticBundle/Resources/config/pimcore/admin.yml
@@ -1,4 +1,0 @@
-core_shop_index:
-    pimcore_admin:
-        js:
-            index_worker_elastic: '/bundles/elasticsearch/pimcore/js/index/worker/elastic.js'
