--- src/CoreShop/Bundle/IndexElasticBundle/Resources/config/pimcore/config.yml
@@ -1,2 +1,0 @@
-imports:
-    - { resource: admin.yml }
