--- src/CoreShop/Bundle/IndexElasticBundle/Resources/config/services/order_renderer.yml
@@ -1,12 +1,0 @@
-services:
-    ## Renderer
-    CoreShop\Bundle\IndexElasticBundle\Order\Elasticsearch\AbstractElasticsearchDynamicRenderer:
-        abstract: true
-        arguments:
-            - '@doctrine.dbal.default_connection'
-
-    CoreShop\Bundle\IndexElasticBundle\Order\Elasticsearch\SimpleOrderRenderer:
-        parent: CoreShop\Bundle\IndexElasticBundle\Order\Elasticsearch\AbstractElasticsearchDynamicRenderer
-        tags:
-            - { name: coreshop.index.order.renderer, type: elastic_simple }
-
