--- src/CoreShop/Bundle/IndexElasticBundle/composer.json
@@ -1,53 +1,0 @@
-{
-  "name": "coreshop/index-elastic-bundle",
-  "type": "pimcore-bundle",
-  "description": "CoreShop - Elasticsearch Bundle",
-  "keywords": [
-    "coreshop",
-    "pimcore",
-    "ecommerce"
-  ],
-  "homepage": "https://www.coreshop.org",
-  "license": "GPL-3.0-or-later",
-  "authors": [
-    {
-      "name": "CoreShop Team",
-      "email": "info@coreshop.org",
-      "homepage": "https://www.coreshop.org/",
-      "role": "Developer"
-    },
-    {
-      "name": "Dominik Pfaffenbauer",
-      "email": "dominik@pfaffenbauer.at",
-      "homepage": "https://www.pfaffenbauer.at/",
-      "role": "Developer"
-    }
-  ],
-  "require": {
-    "coreshop/index-bundle": "^3.0",
-    "elasticsearch/elasticsearch": "^8"
-  },
-  "require-dev": {
-    "phpstan/phpstan": "^1.5.4",
-    "phpstan/phpstan-doctrine": "^1.3.2",
-    "phpstan/phpstan-webmozart-assert": "^1.1.2"
-  },
-  "autoload": {
-    "psr-4": {
-      "CoreShop\\Bundle\\IndexElasticBundle\\": ""
-    }
-  },
-  "repositories": [
-    {
-      "type": "path",
-      "url": "../../*/*"
-    }
-  ],
-  "minimum-stability": "dev",
-  "prefer-stable": true,
-  "extra": {
-    "branch-alias": {
-      "dev-master": "3.0-dev"
-    }
-  }
-}
